+++ nestkernel/mpi_manager.h
@@ -48,8 +48,6 @@
 #include "spike_data.h"
 #include "target_data.h"
 
-// Includes from sli:
-#include "dictdatum.h"
 
 namespace nest
 {
@@ -128,7 +126,7 @@ public:
    * gather all send_buffer vectors on other mpi process to recv_buffer
    * vector
    */
-  void communicate( std::vector< long >& send_buffer, std::vector< long >& recv_buffer );
+  void communicate( std::vector< size_t >& send_buffer, std::vector< size_t >& recv_buffer );
 
   void communicate( std::vector< unsigned int >& send_buffer,
     std::vector< unsigned int >& recv_buffer,
